# ui/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "quarry-server",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.7-nightly.20260731"
  },
  "paths": {
...
    "/v1/libraries/{library}/documents-by-id/{document_id}/document-commands": {
      "post": {
        "tags": [
          "document_engine"
        ],
        "operationId": "commands_by_id",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentCommandAck"
                }
              }
            }
          }
        }
      }
    },
    "/v1/libraries/{library}/documents-by-id/{document_id}/document-state": {
      "get": {
        "tags": [
          "document_engine"
        ],
        "operationId": "read_by_id",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Comma-separated native heads already held by the client",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/libraries/{library}/documents-by-id/{document_id}/events/stream": {
      "get": {
        "tags": [
          "document_engine"
        ],
        "operationId": "events_by_id",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Native document events"
          }
        }
      }
    },
    "/v1/libraries/{library}/documents-by-id/{document_id}/selection": {
      "post": {
        "tags": [
          "document_engine"
        ],
        "operationId": "selection_by_id",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentSelection"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/libraries/{library}/documents-by-id/{document_id}/transactions": {
      "post": {
        "tags": [
          "document_engine"
        ],
        "operationId": "transactions_by_id",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Committed document transaction"
          }
        }
      }
    },
...
    "/v1/libraries/{library}/documents/{path}/archive": {
      "get": {
        "tags": [
          "document_engine"
        ],
        "operationId": "library_archive_get_openapi",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentArchive"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "document_engine"
        ],
        "operationId": "library_archive_post_openapi",
        "parameters": [
          {
            "name": "library",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentArchive"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteOutcome"
                }
              }
            }
          },
          "412": {
            "description": "Destination exists"
          }
        }
      }
    },
...
    "/v1/tmp/documents/{secret}/archive": {
      "get": {
        "tags": [
          "document_engine"
        ],
        "operationId": "tmp_archive_get_openapi",
        "parameters": [
          {
            "name": "secret",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentArchive"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "document_engine"
        ],
        "operationId": "tmp_archive_post_openapi",
        "parameters": [
          {
            "name": "secret",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentArchive"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteOutcome"
                }
              }
            }
          },
          "412": {
            "description": "Destination exists"
          }
        }
      }
    },
...
    "/v1/tmp/documents/{secret}/selection": {
      "post": {
        "tags": [
          "document_engine"
        ],
        "operationId": "tmp_selection_openapi",
        "parameters": [
          {
            "name": "secret",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentSelection"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandBatch": {
        "type": "object",
        "required": [
          "request_id",
          "actor",
          "requests"
        ],
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/DocumentActor"
          },
          "request_id": {
            "type": "string"
          },
          "requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommandRequest"
            }
          }
        },
        "additionalProperties": false
      },
      "CommandRequest": {
        "type": "object",
        "description": "The browser and authority execute this same request against the same base.\nIts actor and generated segment IDs are deterministic. Commands following\nan unacknowledged edit therefore retain valid native character references.",
        "required": [
          "request_id",
          "base",
          "commands"
        ],
        "properties": {
          "at": {
            "type": "string"
          },
          "base": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "commands": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Command"
            }
          },
          "request_id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "DocumentActor": {
        "type": "object",
        "required": [
          "kind"
        ],
        "properties": {
          "id": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "type": "string"
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
      "DocumentArchive": {
        "type": "object",
        "description": "Portable native history, review targets, and document metadata. Markdown\nalone cannot represent retained characters or every review decision.",
        "required": [
          "format",
          "version",
          "bytes"
        ],
        "properties": {
          "bytes": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          "format": {
            "type": "string"
          },
          "metadata": {},
          "version": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "DocumentCommandAck": {
        "type": "object",
        "required": [
          "request_id",
          "document_clock",
          "transaction_id",
          "heads",
          "changed_block_ids"
        ],
        "properties": {
          "changed_block_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "document_clock": {
            "type": "string"
          },
          "heads": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "request_id": {
            "type": "string"
          },
          "transaction_id": {
            "type": "string"
          }
        }
      },
...
      "DocumentId": {
        "type": "string"
      },
...
      "DocumentListEntry": {
        "type": "object",
        "required": [
          "id",
          "path",
          "head_version_id",
          "content_type",
          "byte_size",
          "metadata",
          "updated_at"
        ],
        "properties": {
          "byte_size": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "content_hash": {
            "type": [
              "string",
              "null"
            ]
          },
          "content_type": {
            "type": "string"
          },
          "expires_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Timestamp"
              }
            ]
          },
          "head_version_id": {
            "$ref": "#/components/schemas/VersionId"
          },
          "id": {
            "$ref": "#/components/schemas/DocumentId"
          },
          "library_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {
            "$ref": "#/components/schemas/Value"
          },
          "path": {
            "type": "string"
          },
          "updated_at": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
      "DocumentSelection": {
        "type": "object",
        "required": [
          "client_id",
          "author",
          "points"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TextPoint"
            }
          }
        },
        "additionalProperties": false
      },
...
      "DocumentStateResponse": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "writable",
              "metadata",
              "document_clock",
              "bytes",
              "document",
              "format"
            ],
            "properties": {
              "base": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "bytes": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              },
              "document": {
                "$ref": "#/components/schemas/DocumentView"
              },
              "document_clock": {
                "type": "string"
              },
              "format": {
                "type": "string",
                "enum": [
                  "automerge"
                ]
              },
              "metadata": {},
              "writable": {
                "type": "boolean"
              }
            }
          }
        ]
      },
...
      "DocumentView": {
        "type": "object",
        "required": [
          "document_id",
          "schema_version",
          "heads",
          "blocks",
          "comments",
          "proposals",
          "conflicts"
        ],
        "properties": {
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockView"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentView"
            }
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Conflict"
            }
          },
          "document_id": {
            "type": "string"
          },
          "heads": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "proposals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProposalView"
            }
          },
          "schema_version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "TextPoint": {
        "type": "object",
        "required": [
          "source",
          "cursor"
        ],
        "properties": {
          "cursor": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        }
      },
...
      "Timestamp": {
        "type": "string"
      },
...
      "TransactionRecord": {
        "type": "object",
        "required": [
          "id",
          "library_id",
          "state",
          "source",
          "provenance",
          "created_at"
        ],
        "properties": {
          "actor": {
            "type": [
              "string",
              "null"
            ]
          },
          "committed_at": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Timestamp"
              }
            ]
          },
          "created_at": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "id": {
            "type": "string"
          },
          "library_id": {
            "type": "string"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/Value"
          },
          "source": {
            "$ref": "#/components/schemas/DocumentSource"
          },
          "state": {
            "$ref": "#/components/schemas/TransactionState"
          }
        }
      },
...
      "Value": {},
...
      "VersionId": {
        "type": "string"
      },
      "WriteOutcome": {
        "type": "object",
        "required": [
          "document",
          "version",
          "transaction"
        ],
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentListEntry"
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionRecord"
          },
          "version": {
            "$ref": "#/components/schemas/DocumentVersion"
          }
        }
      }
    }
  }
}